% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownCloud4R.R
\docType{package}
\name{ownCloud4R}
\alias{ownCloud4R}
\alias{ownCloud4R-package}
\title{Interface to 'ownCloud' REST API}
\description{
Provides an Interface to 'Zenodo' (<https://owncloud.org/>) REST API.
}
\details{
\tabular{ll}{
   Package: \tab ownCloud4R\cr
   Type: \tab Package\cr
   Version: \tab 0.0.9000\cr
   Date: \tab 2019-11-06\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
