% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownCloud4R_logger.R
\docType{class}
\name{ownCloud4RLogger}
\alias{ownCloud4RLogger}
\title{ownCloud4RLogger}
\format{\code{\link{R6Class}} object.}
\usage{
ownCloud4RLogger
}
\value{
Object of \code{\link{R6Class}} for modelling a simple logger
}
\description{
ownCloud4RLogger
}
\note{
Logger class used internally by ownCloud4R
}
\section{Abstract Methods}{

\describe{
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
}
}

\keyword{logger}
