% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownCloudManager.R
\docType{class}
\name{ownCloudManager}
\alias{ownCloudManager}
\title{ownCloudManager}
\format{\code{\link{R6Class}} object.}
\usage{
ownCloudManager
}
\value{
Object of \code{\link{R6Class}} for modelling an ownCloudManager
}
\description{
ownCloudManager
}
\section{General Methods}{

\describe{
 \item{\code{new(url, user, pwd, logger)}}{
   This method is used to instantiate an ownCloudManager. The user/pwd are
   mandatory in order to connect to 'ownCloud'. The logger can be either
   NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl 
   http calls logs)
 }
 \item{\code{connect()}}{
   A method to connect to 'ownCloud' and set version/capabilities
 }
 \item{\code{getVersion()}}{
   Get the 'ownCloud' server version
 }
 \item{\code{getCapabilities()}}{
   Get the 'ownCloud' server capabilities
 }
 \item{\code{getWebdavRoot()}}{
   Get the 'ownCloud' WebDAV root URL
 }
}
}

\section{WebDAV methods}{

\describe{
 \item{\code{listFiles(relPath)}}{
   WebDAV method to list folders/files given a relative path. The relative path is set
   to \code{"/"} by default, which corresponds to the root of the 'ownCloud' repository.
 }
 \item{\code{makeCollection(name, relPath)}}{
   WebDAV method to make a collection. By default \code{relPath} is set to \code{"/"} (root).
   The \code{name} is the name of the new collection to be created. The function is recursive
   in the sense that a \code{name} can be provided as relative path of a collection tree 
   (eg \code{newfolder1/newfolder2/newfolder3}), the function will create recursively as 
   many collections are handled in the name. 
 }
}
}

\section{'ownCloud' Share API methods}{

\describe{
 \item{\code{getShares(path, reshares, shared_with_me, state, subfiles, pretty)}}{
   Get list of shares as \code{list}. To return as \code{data.frame}, set 
   \code{pretty = TRUE}. The method accepts additional parameters .More info
   at 
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{manager}
\keyword{owncloud}
